+++ example/test/basic_test.dart
@@ -128,5 +128,13 @@ void main() {
       assert(work.errorType == null);
       assert(work.fromCache);
     });
+
+    test('restart', () async {
+      final work = await RestartWork('超悟空', 32).start();
+
+      print('work result ${work.result} message ${work.message}');
+      assert(work.success);
+      assert(work.errorType == null);
+    });
   });
 }