+++ lib/src/_work_request.dart
@@ -17,22 +17,23 @@ import 'work_model.dart';
 /// [tag]为跟踪日志标签，[options]为请求所需的全部参数，返回响应数据
 Future<HttpCall> workRequest(String tag, WorkRequestOptions options) async {
   final client = (workConfigs[options.configKey] ?? workConfig).dio;
-  final dioOptions = options.toDioOptions();
+  final dioOptions = options.dioOptions;
 
   Map<String, dynamic>? queryParameters = options.queryParams;
 
   dynamic data;
 
-  switch (options.method) {
-    case HttpMethod.get:
-    case HttpMethod.head:
+  switch (dioOptions.method) {
+    case 'GET':
+    case 'HEAD':
       if (options.params is Map<String, dynamic>) {
         queryParameters ??= options.params;
       }
       break;
     default:
-      final isFormData = (options.contentType ?? client.options.contentType) ==
-          multipartFormData;
+      final isFormData =
+          (dioOptions.contentType ?? client.options.contentType) ==
+              multipartFormData;
       data = isFormData && options.params is Map<String, dynamic>
           ? await convertToDio(options.params)
           : options.params;
@@ -65,17 +66,21 @@ extension DioErrorConvert on DioErrorType {
   /// 转换dio异常类型到work库异常类型
   WorkErrorType toWorkErrorType() {
     switch (this) {
-      case DioErrorType.connectTimeout:
-        return WorkErrorType.connectTimeout;
+      case DioErrorType.connectionTimeout:
+        return WorkErrorType.connectionTimeout;
       case DioErrorType.sendTimeout:
         return WorkErrorType.sendTimeout;
       case DioErrorType.receiveTimeout:
         return WorkErrorType.receiveTimeout;
-      case DioErrorType.response:
+      case DioErrorType.badResponse:
         return WorkErrorType.response;
       case DioErrorType.cancel:
         return WorkErrorType.cancel;
-      default:
+      case DioErrorType.badCertificate:
+        return WorkErrorType.certificate;
+      case DioErrorType.connectionError:
+        return WorkErrorType.connection;
+      case DioErrorType.unknown:
         return WorkErrorType.other;
     }
   }
@@ -107,42 +112,7 @@ extension ResponseConvert on Response {
       statusCode: statusCode ?? 0,
       headers: headers.map,
       data: workData,
+      realUri: realUri,
     );
-  }
-}
-
-/// 为[HttpMethod]扩展方法
-extension HttpMethodExtension on HttpMethod {
-  /// 对应的http方法名称
-  String get name {
-    switch (this) {
-      case HttpMethod.get:
-        return 'GET';
-      case HttpMethod.post:
-        return 'POST';
-      case HttpMethod.put:
-        return 'PUT';
-      case HttpMethod.head:
-        return 'HEAD';
-      case HttpMethod.patch:
-        return 'PATCH';
-      case HttpMethod.delete:
-        return 'DELETE';
-    }
-  }
-}
-
-/// [WorkRequestOptions]扩展
-extension WorkRequestOptionsConvert on WorkRequestOptions {
-  /// 转换到dio[Options]
-  Options toDioOptions() {
-    return Options()
-      ..method = method.name
-      ..headers = headers
-      ..contentType = contentType
-      ..receiveTimeout = readTimeout
-      ..sendTimeout = sendTimeout
-      ..responseType = responseType
-      ..listFormat = listFormat;
   }
 }